+++ quil-rs/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.22.1-rc.0
+
+### Fixes
+
+- misc instruction memory accesses (#304)
+
 ## 0.22.0
 
 ### Breaking Changes